add deterministic trainer config and model wrapper
