+++ configs/training/trainer_flow_matching.yaml
