+++ configs/vjepa2_kinetics_400.yaml
@@ -1,5 +1,5 @@
 inherits:
   - datasets/kinetics_400.yaml
   - backbones/vjepa2.yaml
-  - training/trainer.yaml
+  - training/trainer_flow_matching.yaml
 encoder_trainable: false
