+++ configs/vjepa2_kinetics_400_cached.yaml
@@ -1,5 +1,5 @@
 inherits:
   - datasets/kinetics_400_cached.yaml
-  - training/trainer.yaml
+  - training/trainer_flow_matching.yaml
 encoder_trainable: false
 backbone: null
